Update for trac ticket 88 - Native 64bit MapGuide Support http://trac.osgeo.org/mapguide/ticket/88
- Updated project/solution files with x64 configurations for Windows
- Found and fixed x64 portability issue in StringCollection
- Tested server unit tests on Windows XP x64 - all passed
- Tested web tier on Windows XP x64 via the web tier test pages - some passed
- Tried to connect with Studio and MapAdmin to x64 MapGuide server - both failed to connect (investigating further)

Note: There is still more work to do for x64 support in MapGuide

git-svn-id: http://svn.osgeo.org/mapguide/trunk@3415 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Foundation/Data/StringCollection.cpp
@@ -312,12 +312,12 @@ MgStringCollection* MgStringCollection::ParseCollection(CREFSTRING coll, CREFSTR
         return NULL;
 
     STRING separator = delim.length() == 0? L",": delim;
-    unsigned int delimPos, currPos = 0;
+    size_t delimPos, currPos = 0;
     bool done = false;
     STRING elem;
     while(!done)
     {
-        delimPos = (int)collection.find_first_of(separator, currPos);
+        delimPos = collection.find_first_of(separator, currPos);
         if(delimPos == wstring::npos)
         {
             elem = collection.substr(currPos);
